+++ pingone/types/role.go
@@ -1,0 +1,17 @@
+package types
+
+func RoleNameList() []string {
+
+	return []string{"Organization Admin",
+		"Environment Admin",
+		"Identity Data Admin",
+		"Client Application Developer",
+		"Identity Data Read Only",
+		"Configuration Read Only",
+		"PingFederate Administrator",
+		"PingFederate Expression Administrator",
+		"PingFederate Crypto Administrator",
+		"PingFederate User Administrator",
+		"PingFederate Auditor",
+	}
+}