+++ CHANGELOG.md
@@ -1,8 +1,8 @@
-# Release (Unreleased)
+# Release (2022-07-16-2)
 
 * **Feature** `github.com/patrickcping/pingone-go-sdk-v2` : v0.2.0
     * New `pingone` package with aggregated client initialisation
-    * New `environment` and `schemaAttribute` helper types
+    * New `environment`, `schemaAttribute` and `role` helper types
 
 # Release (2022-07-16)